requirements.txt
sentence-transformers
faiss-cpu
gradio
numpy